requirements.txt
discord.py==1.7.3
mysql
mysql_connector_repackaged==0.3.1
